backend/requirements.txt
# Python dependencies for Cori Electron App backend
flask==2.3.3
flask-cors==5.0.1
openai==1.66.3
python-dotenv==1.0.1
requests==2.32.3
# Additional dependencies for core infrastructure
pydantic==2.6.3
pillow==10.2.0
# Windows-specific dependencies (required)
pywin32==306
